requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0